Move getBuildStatusClient and getDeploymentClient off of BitbucketClientFactory and onto BitbucketRepositoryClient
