+++ src/main/java/com/atlassian/bitbucket/jenkins/internal/client/BitbucketClientFactory.java
@@ -1,8 +1,5 @@
 package com.atlassian.bitbucket.jenkins.internal.client;
 
-import com.atlassian.bitbucket.jenkins.internal.model.BitbucketCICapabilities;
-import com.atlassian.bitbucket.jenkins.internal.scm.BitbucketSCMRepository;
-
 /**
  * Factory for Bitbucket Clients.
  */
@@ -22,26 +19,6 @@ public interface BitbucketClientFactory {
      * @return a client that is ready to use
      */
     BitbucketCapabilitiesClient getCapabilityClient();
-
-    /**
-     * Return a client that can post the current status of a build to Bitbucket.
-     *
-     * @param revisionSha      the revision for the build status
-     * @param bitbucketSCMRepo the (Bitbucket) SCM Repo
-     * @param ciCapabilities   CI capabilities of the remote server
-     * @return a client that can post a build status
-     */
-    BitbucketBuildStatusClient getBuildStatusClient(String revisionSha, BitbucketSCMRepository bitbucketSCMRepo,
-                                                    BitbucketCICapabilities ciCapabilities);
-
-    /**
-     * Return a client that can post deployment information to Bitbucket.
-     *
-     * @param revisionSha      the revision for the deployment
-     * @param bitbucketSCMRepo the (Bitbucket) SCM Repo
-     * @return a client that can post deployment information
-     */
-    BitbucketDeploymentClient getDeploymentClient(String revisionSha, BitbucketSCMRepository bitbucketSCMRepo);
 
     /**
      * Construct a client that can retrieve the list of mirrored repositories for a given {@code repoId} from Bitbucket.
