+++ src/main/java/com/atlassian/bitbucket/jenkins/internal/client/BitbucketClientFactoryImpl.java
@@ -2,12 +2,7 @@ package com.atlassian.bitbucket.jenkins.internal.client;
 
 import com.atlassian.bitbucket.jenkins.internal.client.supply.BitbucketCapabilitiesSupplier;
 import com.atlassian.bitbucket.jenkins.internal.credentials.BitbucketCredentials;
-import com.atlassian.bitbucket.jenkins.internal.model.BitbucketCICapabilities;
-import com.atlassian.bitbucket.jenkins.internal.provider.InstanceKeyPairProvider;
-import com.atlassian.bitbucket.jenkins.internal.scm.BitbucketSCMRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.annotations.VisibleForTesting;
-import org.jenkinsci.plugins.displayurlapi.DisplayURLProvider;
 
 public class BitbucketClientFactoryImpl implements BitbucketClientFactory {
 
@@ -29,36 +24,6 @@ public class BitbucketClientFactoryImpl implements BitbucketClientFactory {
     @Override
     public BitbucketCapabilitiesClient getCapabilityClient() {
         return new BitbucketCapabilitiesClientImpl(bitbucketRequestExecutor, capabilitiesSupplier);
-    }
-
-    @VisibleForTesting
-    public BitbucketBuildStatusClient getBuildStatusClient(String revisionSha,
-                                                           BitbucketSCMRepository bitbucketSCMRepo,
-                                                           BitbucketCICapabilities ciCapabilities,
-                                                           InstanceKeyPairProvider instanceKeyPairProvider,
-                                                           DisplayURLProvider displayURLProvider) {
-        if (ciCapabilities.supportsRichBuildStatus()) {
-            return new ModernBitbucketBuildStatusClientImpl(bitbucketRequestExecutor, bitbucketSCMRepo.getProjectKey(),
-                    bitbucketSCMRepo.getRepositorySlug(), revisionSha, instanceKeyPairProvider, displayURLProvider);
-        }
-        return new BitbucketBuildStatusClientImpl(bitbucketRequestExecutor, revisionSha);
-    }
-
-    @Override
-    public BitbucketBuildStatusClient getBuildStatusClient(String revisionSha,
-                                                           BitbucketSCMRepository bitbucketSCMRepo,
-                                                           BitbucketCICapabilities ciCapabilities) {
-        if (ciCapabilities.supportsRichBuildStatus()) {
-            return new ModernBitbucketBuildStatusClientImpl(bitbucketRequestExecutor, bitbucketSCMRepo.getProjectKey(),
-                    bitbucketSCMRepo.getRepositorySlug(), revisionSha);
-        }
-        return new BitbucketBuildStatusClientImpl(bitbucketRequestExecutor, revisionSha);
-    }
-
-    @Override
-    public BitbucketDeploymentClient getDeploymentClient(String revisionSha, BitbucketSCMRepository bitbucketSCMRepo) {
-        return new BitbucketDeploymentClientImpl(bitbucketRequestExecutor, bitbucketSCMRepo.getProjectKey(),
-                bitbucketSCMRepo.getRepositorySlug(), revisionSha);
     }
 
     @Override
